Made podcast creation/editing work.

+++ Teknik/Areas/Podcast/Controllers/PodcastController.cs
@@ -1,9 +1,11 @@
 ﻿using System;
 using System.Collections.Generic;
 using System.Data.Entity;
+using System.IO;
 using System.Linq;
 using System.Web;
 using System.Web.Mvc;
+using Teknik.Areas.Podcast.Models;
 using Teknik.Areas.Podcast.ViewModels;
 using Teknik.Controllers;
 using Teknik.Models;
@@ -13,21 +15,19 @@ namespace Teknik.Areas.Podcast.Controllers
     public class PodcastController : DefaultController
     {
         private TeknikEntities db = new TeknikEntities();
-
-        // GET: Blogs/Details/5
+        
         [AllowAnonymous]
         public ActionResult Index()
         {
             MainViewModel model = new MainViewModel();
+            model.Title = Config.PodcastConfig.Title;
+            model.Description = Config.PodcastConfig.Description;
             try
             {
-                ViewBag.Title = "Teknik Blog - " + Config.Title;
-                var foundPodcasts = (User.IsInRole("Admin")) ? db.Podcasts.FirstOrDefault() : db.Podcasts.Where(p => (p.Published)).FirstOrDefault();
+                ViewBag.Title = "Teknikast - " + Config.Title;
+                var foundPodcasts = (User.IsInRole("Podcast")) ? db.Podcasts.FirstOrDefault() : db.Podcasts.Where(p => (p.Published)).FirstOrDefault();
                 if (foundPodcasts != null)
                 {
-                    model = new MainViewModel();
-                    model.Title = Config.PodcastConfig.Title;
-                    model.Description = Config.PodcastConfig.Description;
                     model.HasPodcasts = (foundPodcasts != null);
                 }
                 else
@@ -47,14 +47,13 @@ namespace Teknik.Areas.Podcast.Controllers
             }
         }
 
-        #region Posts
-        // GET: Blogs/Details/5
+        #region Podcasts
         [AllowAnonymous]
         public ActionResult View(int episode)
         {
             PodcastViewModel model = new PodcastViewModel();
             // find the podcast specified
-            var foundPodcast = (User.IsInRole("Admin")) ? db.Podcasts.Where(p => p.Episode == episode).FirstOrDefault() : db.Podcasts.Where(p => (p.Published && p.Episode == episode)).FirstOrDefault();
+            var foundPodcast = (User.IsInRole("Podcast")) ? db.Podcasts.Include("Files").Where(p => p.Episode == episode).FirstOrDefault() : db.Podcasts.Include("Files").Where(p => (p.Published && p.Episode == episode)).FirstOrDefault();
             if (foundPodcast != null)
             {
                 model.PodcastId = foundPodcast.PodcastId;
@@ -71,12 +70,43 @@ namespace Teknik.Areas.Podcast.Controllers
             return View("~/Areas/Podcast/Views/Podcast/ViewPodcast.cshtml", model);
         }
 
+        [AllowAnonymous]
+        public ActionResult Download(int episode, string fileName)
+        {
+            // find the podcast specified
+            var foundPodcast = (User.IsInRole("Podcast")) ? db.Podcasts.Include("Files").Where(p => p.Episode == episode).FirstOrDefault() : db.Podcasts.Include("Files").Where(p => (p.Published && p.Episode == episode)).FirstOrDefault();
+            if (foundPodcast != null)
+            {
+                PodcastFile file = foundPodcast.Files.Where(f => f.FileName == fileName).FirstOrDefault();
+                if (file != null)
+                {
+                    if (System.IO.File.Exists(file.Path))
+                    {
+                        // Read in the file
+                        byte[] data = System.IO.File.ReadAllBytes(file.Path);
+
+                        // Create File
+                        var cd = new System.Net.Mime.ContentDisposition
+                        {
+                            FileName = file.FileName,
+                            Inline = true
+                        };
+
+                        Response.AppendHeader("Content-Disposition", cd.ToString());
+
+                        return File(data, file.ContentType);
+                    }
+                }
+            }
+            return Redirect(Url.SubRouteUrl("error", "Error.Http404"));
+        }
+
         [HttpPost]
         [AllowAnonymous]
         public ActionResult GetPodcasts(int startPodcastID, int count)
         {
-            var podcasts = (User.IsInRole("Admin")) ? db.Podcasts.OrderByDescending(p => p.DatePosted).Skip(startPodcastID).Take(count).ToList()
-                                                    : db.Podcasts.Where(p => p.Published).OrderByDescending(p => p.DatePosted).Skip(startPodcastID).Take(count).ToList();
+            var podcasts = (User.IsInRole("Podcast")) ? db.Podcasts.Include("Files").OrderByDescending(p => p.DatePosted).Skip(startPodcastID).Take(count).ToList()
+                                                    : db.Podcasts.Include("Files").Where(p => p.Published).OrderByDescending(p => p.DatePosted).Skip(startPodcastID).Take(count).ToList();
             List<PodcastViewModel> podcastViews = new List<PodcastViewModel>();
             if (podcasts != null)
             {
@@ -90,65 +120,133 @@ namespace Teknik.Areas.Podcast.Controllers
 
         [HttpPost]
         [AllowAnonymous]
+        public ActionResult GetPodcastEpisode(int podcastId)
+        {
+            var foundPodcast = (User.IsInRole("Podcast")) ? db.Podcasts.Where(p => p.PodcastId == podcastId).FirstOrDefault() : db.Podcasts.Where(p => (p.Published && p.PodcastId == podcastId)).FirstOrDefault();
+            if (foundPodcast != null)
+            {
+                return Json(new { result = foundPodcast.Episode });
+            }
+            return Json(new { error = "No podcast found" });
+        }
+
+        [HttpPost]
+        [AllowAnonymous]
         public ActionResult GetPodcastTitle(int podcastId)
         {
-            var foundPodcast = (User.IsInRole("Admin")) ? db.Podcasts.Where(p => p.PodcastId == podcastId).FirstOrDefault() : db.Podcasts.Where(p => (p.Published && p.PodcastId == podcastId)).FirstOrDefault();
+            var foundPodcast = (User.IsInRole("Podcast")) ? db.Podcasts.Where(p => p.PodcastId == podcastId).FirstOrDefault() : db.Podcasts.Where(p => (p.Published && p.PodcastId == podcastId)).FirstOrDefault();
             if (foundPodcast != null)
             {
                 return Json(new { result = foundPodcast.Title });
             }
-            return Json(new { error = "No title found" });
+            return Json(new { error = "No podcast found" });
         }
 
         [HttpPost]
         [AllowAnonymous]
         public ActionResult GetPodcastDescription(int podcastId)
         {
-            var foundPodcast = (User.IsInRole("Admin")) ? db.Podcasts.Where(p => p.PodcastId == podcastId).FirstOrDefault() : db.Podcasts.Where(p => (p.Published && p.PodcastId == podcastId)).FirstOrDefault();
+            var foundPodcast = (User.IsInRole("Podcast")) ? db.Podcasts.Where(p => p.PodcastId == podcastId).FirstOrDefault() : db.Podcasts.Where(p => (p.Published && p.PodcastId == podcastId)).FirstOrDefault();
             if (foundPodcast != null)
             {
                 return Json(new { result = foundPodcast.Description });
             }
-            return Json(new { error = "No article found" });
+            return Json(new { error = "No podcast found" });
         }
 
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult CreatePodcast(string title, string description, HttpPostedFileBase[] files)
+        public ActionResult CreatePodcast(int episode, string title, string description)
         {
             if (ModelState.IsValid)
             {
-                // Handle saving of files
-                Models.Podcast podcast = db.Podcasts.Create();
-                podcast.Title = title;
-                podcast.Description = description;
-                podcast.DatePosted = DateTime.Now;
-                podcast.DatePublished = DateTime.Now;
+                if (User.IsInRole("Podcast"))
+                {
+                    // Grab the next episode number
+                    Models.Podcast lastPod = db.Podcasts.Where(p => p.Episode == episode).FirstOrDefault();
+                    if (lastPod == null)
+                    {
+                        if (Request.Files.Count > 0)
+                        {
+                            // Create the podcast object
+                            Models.Podcast podcast = db.Podcasts.Create();
+                            podcast.Episode = episode;
+                            podcast.Title = title;
+                            podcast.Description = description;
+                            podcast.DatePosted = DateTime.Now;
+                            podcast.DatePublished = DateTime.Now;
 
-                db.Podcasts.Add(podcast);
-                db.SaveChanges();
-                return Json(new { result = true });
+                            // Handle saving of files
+                            for (int i = 0; i < Request.Files.Count; i++)
+                            {
+                                HttpPostedFileBase file = Request.Files[i]; //Uploaded file
+                                                                            //Use the following properties to get file's name, size and MIMEType
+                                int fileSize = file.ContentLength;
+                                string fileName = file.FileName;
+                                string mimeType = file.ContentType;
+                                string fileExt = Path.GetExtension(fileName);
+                                if (!Directory.Exists(Config.PodcastConfig.PodcastDirectory))
+                                {
+                                    Directory.CreateDirectory(Config.PodcastConfig.PodcastDirectory);
+                                }
+                                string newName = string.Format("Teknikast_Episode_{0}{1}", episode, fileExt);
+                                int index = 1;
+                                while (System.IO.File.Exists(Path.Combine(Config.PodcastConfig.PodcastDirectory, newName)))
+                                {
+                                    newName = string.Format("Teknikast_Episode_{0} ({1}){2}", episode, index, fileExt);
+                                    index++;
+                                }
+                                string fullPath = Path.Combine(Config.PodcastConfig.PodcastDirectory, newName);
+                                PodcastFile podFile = new PodcastFile();
+                                podFile.Path = fullPath;
+                                podFile.FileName = newName;
+                                podFile.ContentType = mimeType;
+                                podcast.Files = new List<PodcastFile>();
+                                podcast.Files.Add(podFile);
+
+                                file.SaveAs(fullPath);
+                            }
+
+                            db.Podcasts.Add(podcast);
+                            db.SaveChanges();
+                            return Json(new { result = true });
+                        }
+                        return Json(new { error = "You must submit at least one podcast audio file" });
+                    }
+                    return Json(new { error = "That episode already exists" });
+                }
+                return Json(new { error = "You don't have permission to create a podcast" });
             }
             return Json(new { error = "No podcast created" });
         }
 
         [HttpPost]
         [ValidateAntiForgeryToken]
-        public ActionResult EditPodcast(int podcastId, string title, string description)
+        public ActionResult EditPodcast(int podcastId, int episode, string title, string description)
         {
             if (ModelState.IsValid)
             {
-                Models.Podcast podcast = db.Podcasts.Find(podcastId);
-                if (podcast != null)
+                if (User.IsInRole("Podcast"))
                 {
-                    podcast.Title = title;
-                    podcast.Description = description;
-                    db.Entry(podcast).State = EntityState.Modified;
-                    db.SaveChanges();
-                    return Json(new { result = true });
+                    Models.Podcast podcast = db.Podcasts.Find(podcastId);
+                    if (podcast != null)
+                    {
+                        if (db.Podcasts.Where(p => p.Episode != episode).FirstOrDefault() == null)
+                        {
+                            podcast.Episode = episode;
+                            podcast.Title = title;
+                            podcast.Description = description;
+                            db.Entry(podcast).State = EntityState.Modified;
+                            db.SaveChanges();
+                            return Json(new { result = true });
+                        }
+                        return Json(new { error = "That episode already exists" });
+                    }
+                    return Json(new { error = "No podcast found" });
                 }
+                return Json(new { error = "You don't have permission to edit this podcast" });
             }
-            return Json(new { error = "No podcast found" });
+            return Json(new { error = "Invalid Inputs" });
         }
 
         [HttpPost]
@@ -157,18 +255,23 @@ namespace Teknik.Areas.Podcast.Controllers
         {
             if (ModelState.IsValid)
             {
-                Models.Podcast podcast = db.Podcasts.Find(podcastId);
-                if (podcast != null)
+                if (User.IsInRole("Podcast"))
                 {
-                    podcast.Published = publish;
-                    if (publish)
-                        podcast.DatePublished = DateTime.Now;
-                    db.Entry(podcast).State = EntityState.Modified;
-                    db.SaveChanges();
-                    return Json(new { result = true });
+                    Models.Podcast podcast = db.Podcasts.Find(podcastId);
+                    if (podcast != null)
+                    {
+                        podcast.Published = publish;
+                        if (publish)
+                            podcast.DatePublished = DateTime.Now;
+                        db.Entry(podcast).State = EntityState.Modified;
+                        db.SaveChanges();
+                        return Json(new { result = true });
+                    }
+                    return Json(new { error = "No podcast found" });
                 }
+                return Json(new { error = "You don't have permission to publish this podcast" });
             }
-            return Json(new { error = "No podcast found" });
+            return Json(new { error = "Invalid Inputs" });
         }
 
         [HttpPost]
@@ -177,15 +280,20 @@ namespace Teknik.Areas.Podcast.Controllers
         {
             if (ModelState.IsValid)
             {
-                Models.Podcast podcast = db.Podcasts.Find(podcastId);
-                if (podcast != null)
+                if (User.IsInRole("Podcast"))
                 {
-                    db.Podcasts.Remove(podcast);
-                    db.SaveChanges();
-                    return Json(new { result = true });
+                    Models.Podcast podcast = db.Podcasts.Find(podcastId);
+                    if (podcast != null)
+                    {
+                        db.Podcasts.Remove(podcast);
+                        db.SaveChanges();
+                        return Json(new { result = true });
+                    }
+                    return Json(new { error = "No podcast found" });
                 }
+                return Json(new { error = "You don't have permission to delete this podcast" });
             }
-            return Json(new { error = "No post found" });
+            return Json(new { error = "Invalid Inputs" });
         }
         #endregion
 
@@ -224,17 +332,21 @@ namespace Teknik.Areas.Podcast.Controllers
         {
             if (ModelState.IsValid)
             {
-                Models.PodcastComment comment = db.PodcastComments.Create();
-                comment.PodcastId = podcastId;
-                comment.UserId = db.Users.Where(u => u.Username == User.Identity.Name).First().UserId;
-                comment.Article = article;
-                comment.DatePosted = DateTime.Now;
+                if (db.Podcasts.Where(p => p.PodcastId == podcastId).FirstOrDefault() != null)
+                {
+                    Models.PodcastComment comment = db.PodcastComments.Create();
+                    comment.PodcastId = podcastId;
+                    comment.UserId = db.Users.Where(u => u.Username == User.Identity.Name).First().UserId;
+                    comment.Article = article;
+                    comment.DatePosted = DateTime.Now;
 
-                db.PodcastComments.Add(comment);
-                db.SaveChanges();
-                return Json(new { result = true });
+                    db.PodcastComments.Add(comment);
+                    db.SaveChanges();
+                    return Json(new { result = true });
+                }
+                return Json(new { error = "That podcast does not exist" });
             }
-            return Json(new { error = "No comment created" });
+            return Json(new { error = "Invalid Parameters" });
         }
 
         [HttpPost]
@@ -246,13 +358,18 @@ namespace Teknik.Areas.Podcast.Controllers
                 Models.PodcastComment comment = db.PodcastComments.Find(commentID);
                 if (comment != null)
                 {
-                    comment.Article = article;
-                    db.Entry(comment).State = EntityState.Modified;
-                    db.SaveChanges();
-                    return Json(new { result = true });
+                    if (comment.User.Username == User.Identity.Name || User.IsInRole("Admin"))
+                    {
+                        comment.Article = article;
+                        db.Entry(comment).State = EntityState.Modified;
+                        db.SaveChanges();
+                        return Json(new { result = true });
+                    }
+                    return Json(new { error = "You don't have permission to edit this comment" });
                 }
+                return Json(new { error = "No comment found" });
             }
-            return Json(new { error = "No comment found" });
+            return Json(new { error = "Invalid Parameters" });
         }
 
         [HttpPost]
@@ -264,12 +381,17 @@ namespace Teknik.Areas.Podcast.Controllers
                 Models.PodcastComment comment = db.PodcastComments.Find(commentID);
                 if (comment != null)
                 {
-                    db.PodcastComments.Remove(comment);
-                    db.SaveChanges();
-                    return Json(new { result = true });
+                    if (comment.User.Username == User.Identity.Name || User.IsInRole("Admin"))
+                    {
+                        db.PodcastComments.Remove(comment);
+                        db.SaveChanges();
+                        return Json(new { result = true });
+                    }
+                    return Json(new { error = "You don't have permission to delete this comment" });
                 }
+                return Json(new { error = "No comment found" });
             }
-            return Json(new { error = "No comment found" });
+            return Json(new { error = "Invalid Parameters" });
         }
         #endregion
     }
